Adjust redo of set virt size in set_virt_size_with_min.
We normally check to see if we need to change virtual
size again after we set it once, to check that a
change in scrollbar presence didn't change the
client area, thus necessitating changing the
virtual size.

Now we check if the actual virtual size needs
changing, not just if the client area has changed.