Fix of right,left gap rect draw bug.
If the image is too small for the window,
we draw rectangles on the left and right
sides.  This fixes a bug where the height
of the rectangle was too small.  This
came about because the rectangles can
be positioned a little off-window
to stay on pixel boundary, and in this
case the rectangles need to be a little
larger than the window height.  Previously
the rectangles were strictly as large as
the window height, which in some cases
meant that the bottom of the rectangles
didn't cover the area in need of refresh.